refactor(studio): share the logs rewrite flow, assemble prompts server-side
Addresses code review on b626b3b1f1.

Deduplicate the rewrite orchestration. The SQL editor banner and the Logs
Explorer each implemented the same sequence — read the live SQL, discover
the source's log_attributes keys, call the completion route, guard against
edits that landed mid-request, turn the result into a diff proposal — and
had already drifted: only the banner detected a no-op rewrite, they sourced
keys differently, and the Explorer formatted errors with an `as Error` cast
that would lie for a non-Error throw. Both now call useLegacyLogsRewrite,
which owns the request and the offer's state machine, so the Explorer picks
up no-op detection and typed error extraction. The presentational
LegacyLogsRewriteAdmonition covers every state of that machine, giving both
surfaces the same failure and no-rewrite-needed UI instead of toasts.

Assemble the completion prompt in one place. The route built a
schema/code/instruction message for Postgres but forwarded `prompt`
verbatim for ClickHouse, so the client hand-built an equivalent string.
That left two independent descriptions of the same logs table — one in the
system prompt, one injected client-side — and made
completionMetadata.textBeforeCursor/selection/textAfterCursor mean
different things per dialect. The route now uses one template for both,
branching only the schema section and (for intent: 'rewrite') the
instruction, and lib/ai/clickhouse-logs.ts is the single home for
ClickHouse-logs prompt content. buildLogsCompletionPrompt and
buildClickhouseRewritePrompt are gone; the "no code fences" contract is
stated once in the system prompt rather than three times.

The dialect split had also left inline edits ungrounded: availableKeys was
a parameter no inline call site passed, so an inline edit got weaker
grounding on real log_attributes keys than a full rewrite. Keys now travel
in completionMetadata and both paths supply them.

Move logs-sql-rewrite out of Settings/Logs. Two of its three consumers now
live outside that folder, so the import path was becoming misleading: the
dialect helpers and the fetch wrapper go to data/logs, the shared hook to
hooks/analytics, and the prompt text server-side to lib/ai.

One behavior note: the Explorer's persisted dismissal is not used to seed
the machine, because useLocalStorage reads as false during SSR and would
resurrect a dismissed banner on hydration. It gates the offer at render
instead, while an outcome the user just triggered still shows — a rewrite
started from the toolbar action has to report a failure somewhere.
