+++ apps/studio/data/logs/logs-sql-rewrite.test.ts
@@ -1,91 +1,12 @@
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
 import {
-  buildClickhouseRewritePrompt,
-  buildLogsCompletionPrompt,
   detectLogSource,
   looksLikeLegacyLogsQuery,
   rewriteLogsSqlWithAI,
   shouldOfferLegacyLogsRewrite,
   stripSqlCodeFences,
 } from './logs-sql-rewrite'
-
-describe('buildClickhouseRewritePrompt', () => {
-  it('includes the query, the schema, and a reply-with-only-SQL instruction', () => {
-    const prompt = buildClickhouseRewritePrompt('select count(*) from edge_logs')
-    expect(prompt).toContain('select count(*) from edge_logs')
-    expect(prompt).toContain('log_attributes')
-    expect(prompt).toContain("source = 'edge_logs'")
-    expect(prompt.toLowerCase()).toContain('reply with only')
-  })
-
-  it('demands a rewrite explicitly, since the system prompt only covers ClickHouse SQL generally', () => {
-    const prompt = buildClickhouseRewritePrompt('select 1 from edge_logs')
-    expect(prompt).toContain('REWRITE')
-    expect(prompt).toContain('BigQuery query to rewrite:')
-  })
-
-  it('spells out the FROM-to-logs conversion and shows a worked example', () => {
-    const prompt = buildClickhouseRewritePrompt('select 1 from postgres_logs')
-    expect(prompt).toContain("from logs where source = 'postgres_logs'")
-    expect(prompt.toLowerCase()).toContain('remove every')
-    expect(prompt).toContain('cross join unnest')
-    expect(prompt).toContain('BigQuery:')
-    expect(prompt).toContain('ClickHouse:')
-    expect(prompt).toContain("log_attributes['parsed.error_severity']")
-  })
-
-  it('lists the real log_attributes keys when provided and demands exact paths', () => {
-    const prompt = buildClickhouseRewritePrompt('select 1 from edge_logs', [
-      'request.headers.x_real_ip',
-      'request.cf.country',
-    ])
-    expect(prompt).toContain("log_attributes['request.headers.x_real_ip']")
-    expect(prompt).toContain("log_attributes['request.cf.country']")
-    expect(prompt.toLowerCase()).toContain('exact')
-  })
-
-  it('omits the keys section when none are provided', () => {
-    const prompt = buildClickhouseRewritePrompt('select 1 from edge_logs')
-    expect(prompt).not.toContain('actual log_attributes keys present')
-  })
-})
-
-describe('buildLogsCompletionPrompt', () => {
-  const context = {
-    instruction: 'only keep 5xx responses',
-    textBeforeCursor: "select timestamp from logs where source = 'edge_logs'\n",
-    selection: 'limit 5',
-    textAfterCursor: '\n',
-  }
-
-  it('carries the logs schema, the query context, and the instruction', () => {
-    const prompt = buildLogsCompletionPrompt(context)
-    expect(prompt).toContain('log_attributes (Map(String, String))')
-    expect(prompt).toContain("select timestamp from logs where source = 'edge_logs'")
-    expect(prompt).toContain('<selection>limit 5</selection>')
-    expect(prompt).toContain('only keep 5xx responses')
-  })
-
-  it('asks for only the SQL replacing the selection, with no fences', () => {
-    const prompt = buildLogsCompletionPrompt(context)
-    expect(prompt).toContain('Reply with ONLY the SQL that replaces the <selection> block')
-    expect(prompt).toContain('no markdown code fences')
-  })
-
-  it('marks an empty selection as a cursor position', () => {
-    const prompt = buildLogsCompletionPrompt({ ...context, selection: '' })
-    expect(prompt).toContain('<selection></selection>')
-    expect(prompt).toContain('An empty selection means')
-  })
-
-  it('lists the real log_attributes keys when provided and omits the section otherwise', () => {
-    expect(
-      buildLogsCompletionPrompt({ ...context, availableKeys: ['response.status_code'] })
-    ).toContain("log_attributes['response.status_code']")
-    expect(buildLogsCompletionPrompt(context)).not.toContain('actual log_attributes keys present')
-  })
-})
 
 describe('shouldOfferLegacyLogsRewrite', () => {
   const legacySql = 'select 1 from edge_logs cross join unnest(metadata) as m'
@@ -182,7 +103,7 @@ describe('rewriteLogsSqlWithAI', () => {
     vi.unstubAllGlobals()
   })
 
-  it('posts to the completion endpoint and returns the cleaned query', async () => {
+  it('declares the rewrite intent and sends the query as the selection', async () => {
     const fetchMock = vi.fn().mockResolvedValue({
       ok: true,
       json: async () => '```sql\nselect 1 from logs\n```',
@@ -192,6 +113,7 @@ describe('rewriteLogsSqlWithAI', () => {
     const result = await rewriteLogsSqlWithAI({
       sql: 'select 1 from edge_logs',
       projectRef: 'abc',
+      availableKeys: ['request.method'],
     })
 
     expect(result).toBe('select 1 from logs')
@@ -199,8 +121,14 @@ describe('rewriteLogsSqlWithAI', () => {
     expect(url).toContain('/api/ai/code/complete')
     const body = JSON.parse(init.body)
     expect(body.dialect).toBe('clickhouse')
+    expect(body.intent).toBe('rewrite')
+    // The whole query is the selection, so the rewrite replaces all of it.
     expect(body.completionMetadata.selection).toBe('select 1 from edge_logs')
-    expect(body.completionMetadata.prompt.toLowerCase()).toContain('reply with only')
+    expect(body.completionMetadata.textBeforeCursor).toBe('')
+    expect(body.completionMetadata.textAfterCursor).toBe('')
+    expect(body.completionMetadata.availableKeys).toEqual(['request.method'])
+    // No prompt text is carried client-side — the route owns the instruction.
+    expect(body.completionMetadata.prompt).toBe('')
   })
 
   it('throws when the request fails', async () => {
